+++ integration/airflow/openlineage/airflow/extractors/manager.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Optional
+from typing import Optional, Type
 
 from openlineage.airflow.extractors import TaskMetadata, BaseExtractor, Extractors
 from openlineage.airflow.utils import get_job_name
@@ -9,16 +9,27 @@ class ExtractorManager:
     """Class abstracting management of custom extractors."""
     def __init__(self):
         self.extractors = {}
-        self.extractor_mapper = Extractors()
+        self.task_to_extractor = Extractors()
         self.log = logging.getLogger()
 
-    def extract_metadata(self, dagrun, task, complete: bool, task_instance=None) -> TaskMetadata:
+    def add_extractor(self, operator, extractor: Type[BaseExtractor]):
+        self.task_to_extractor.add_extractor(operator, extractor)
+
+    def extract_metadata(
+        self,
+        dagrun,
+        task,
+        complete: bool = False,
+        task_instance=None
+    ) -> TaskMetadata:
         extractor = self._get_extractor(task)
         task_info = f'task_type={task.__class__.__name__} ' \
             f'airflow_dag_id={task.dag_id} ' \
             f'task_id={task.task_id} ' \
             f'airflow_run_id={dagrun.run_id} '
         if extractor:
+            # Extracting advanced metadata is only possible when extractor for particular operator
+            # is defined. Without it, we can't extract any input or output data.
             try:
                 self.log.debug(
                     f'Using extractor {extractor.__class__.__name__} {task_info}')
@@ -48,7 +59,7 @@ class ExtractorManager:
     def _get_extractor(self, task) -> Optional[BaseExtractor]:
         if task.task_id in self.extractors:
             return self.extractors[task.task_id]
-        extractor = self.extractor_mapper.get_extractor_class(task.__class__)
+        extractor = self.task_to_extractor.get_extractor_class(task.__class__)
         self.log.debug(f'extractor for {task.__class__} is {extractor}')
         if extractor:
             self.extractors[task.task_id] = extractor(task)